Some .sql files
buildTables.sql builds the 3 tables in the DB
dropTables.sql gets rid of all the tables
samplePopulate puts in 3 sample questions with 4 choice and 7 responses each.
Currently they only work with SQLite because in the table definition, SQLite uses AUTOINCREMENT and all(?) other flavors use AUTO_INCREMENT so that may have to be changed later.
When using SQLITE, just make sure you are in the DB and the use ".read filename.sql" and it will execute it. :100: