Fized modal on alter questions
I created my own css file because all of it is unique to my own page. I left mt JS in the HTML file because if I put it all in main.js, we would end up with tons of the same JS since we are all using W3 schools stuff.